requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
python-dotenv
passlib[bcrypt]
python-jose
groq
python-dotenv
passlib
python-jose
